Renamed 'client/src/module_bindings/' to 'client/src/server/bindings/'
