+++ client/src/lib.rs
@@ -1,5 +1,5 @@
 pub mod gameplay;
-pub mod module_bindings;
 pub mod oidc_auth_helper;
+pub mod server;
 pub mod shader;
 pub mod stdb;
