+++ client/src/stdb/connector/subscriptions.rs
@@ -1,6 +1,6 @@
 use spacetimedb_sdk::*;
 
-use crate::module_bindings::*;
+use crate::server::bindings::*;
 
 ////////////////////////////////////////////////////////////////////////////////////////////////
 ////////////////////////////////////////////////////////////////////////////////////////////////
