+++ client/src/stdb/mod.rs
@@ -1,4 +1,2 @@
 pub mod connector;
-pub mod impls;
-pub mod stdsl_id_impls;
 pub mod utils;
